requirements.txt
aerich==0.9.2
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
aiosqlite==0.21.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
asyncclick==8.3.0.7
attrs==25.4.0
bcrypt==5.0.0
certifi==2025.11.12
click==8.3.1
deprecated==1.3.1
dictdiffer==0.9.0
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.125.0
fastapi-cli==0.0.16
fastapi-cloud-cli==0.7.0
fastar==0.8.0
frozenlist==1.8.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
iso8601==2.1.0
jinja2==3.1.6
limits==5.6.0
markdown-it-py==4.0.0
markupsafe==3.0.3
mdurl==0.1.2
multidict==6.7.0
nodeenv==1.9.1
nodejs-wheel-binaries==24.12.0
packaging==25.0
propcache==0.4.1
pydantic==2.12.5
pydantic-core==2.41.5
pygments==2.19.2
pyjwt==2.10.1
pypika-tortoise==0.6.3
pyright==1.1.407
python-dotenv==1.2.1
python-multipart==0.0.21
pytubefix==10.3.6
pytz==2025.2
pyyaml==6.0.3
rich==14.2.0
rich-toolkit==0.17.1
rignore==0.7.6
sentry-sdk==2.48.0
shellingham==1.5.4
slowapi==0.1.9
starlette==0.50.0
tortoise-orm==0.25.2
typer==0.20.0
typing-extensions==4.15.0
typing-inspection==0.4.2
urllib3==2.6.2
uvicorn==0.38.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==15.0.1
wrapt==2.0.1
yarl==1.22.0
